requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.7.8
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyMuPDF==1.23.8
python-multipart==0.0.6
sqlalchemy==1.4.53
databases[sqlite]==0.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
torch==2.0.1
huggingface-hub==0.16.4
transformers==4.31.0